+++ backend/src/main/java/com/projetCloud/projetCloud/repository/signalement/SignalementRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import org.springframework.data.repository.query.Param;
 
-public interface SignalementRepository extends JpaRepository<Signalement, Long> {
+public interface SignalementRepository extends JpaRepository<Signalement, Integer> {
     @Query(value = "SELECT nombre_points, total_surface, total_budget, avancement_pourcent FROM vue_recapitulatif_signalement", nativeQuery = true)
     List<Object[]> getRecapitulatifRaw();
 
@@ -17,6 +17,6 @@ public interface SignalementRepository extends JpaRepository<Signalement, Long> 
     List<Object[]> getSignalementCplRaw();
 
     @Query(value = "SELECT id, titre, description, statut, latitude, longitude, surface_m2, budget, id_entreprise, entreprise, entreprise_adresse, entreprise_contact, id_utilisateur, utilisateur_nom, utilisateur_prenom, utilisateur_email, date_creation FROM vue_infos_signalement WHERE id = :id", nativeQuery = true)
-    List<Object[]> getSignalementCplByIdRaw(@Param("id") Long id);
+    List<Object[]> getSignalementCplByIdRaw(@Param("id") Integer id);
 
 }
